[Store][bind] add new support data types
